+++ axelor-account/src/main/java/com/axelor/apps/account/service/invoice/print/InvoicePrintService.java
@@ -37,7 +37,7 @@ public interface InvoicePrintService {
    * @return the file link to the printed invoice.
    * @throws AxelorException
    */
-  String printInvoice(Invoice invoice, boolean forceRefresh, String format, Integer reportType)
+  String printInvoice(Invoice invoice, boolean forceRefresh, String format, Integer reportType, String locale)
       throws AxelorException, IOException;
 
   /**
@@ -51,7 +51,7 @@ public interface InvoicePrintService {
    * @throws AxelorException
    * @throws IOException
    */
-  File printCopiesToFile(Invoice invoice, boolean forceRefresh, Integer reportType)
+  File printCopiesToFile(Invoice invoice, boolean forceRefresh, Integer reportType, String locale)
       throws AxelorException, IOException;
 
   /**
@@ -64,7 +64,7 @@ public interface InvoicePrintService {
    *     changes made to invoice between last print and current state won't appear on printed copy.
    * @return a file with the invoice as PDF.
    */
-  File getPrintedInvoice(Invoice invoice, boolean forceRefresh, Integer reportType)
+  File getPrintedInvoice(Invoice invoice, boolean forceRefresh, Integer reportType, String locale)
       throws AxelorException;
 
   /**
@@ -74,7 +74,7 @@ public interface InvoicePrintService {
    * @return a file of the invoice printing.
    * @throws AxelorException
    */
-  public File print(Invoice invoice, Integer reportType) throws AxelorException;
+  public File print(Invoice invoice, Integer reportType, String locale) throws AxelorException;
 
   /**
    * Print an invoice, then save the generated file in this invoice.
@@ -83,7 +83,7 @@ public interface InvoicePrintService {
    * @return a metafile of the invoice printing.
    * @throws AxelorException
    */
-  File printAndSave(Invoice invoice, Integer reportType) throws AxelorException;
+  File printAndSave(Invoice invoice, Integer reportType, String locale) throws AxelorException;
 
   /**
    * Print a list of invoices in the same output.
@@ -102,5 +102,5 @@ public interface InvoicePrintService {
    * @return the report settings to print the given invoice
    * @throws AxelorException
    */
-  ReportSettings prepareReportSettings(Invoice invoice, Integer reportType) throws AxelorException;
+  ReportSettings prepareReportSettings(Invoice invoice, Integer reportType, String locale) throws AxelorException;
 }