+++ axelor-business-project/src/main/java/com/axelor/apps/businessproject/service/InvoicingProjectService.java
@@ -477,7 +477,7 @@ public class InvoicingProjectService {
       MetaFiles metaFiles = Beans.get(MetaFiles.class);
 
       fileList.add(
-          Beans.get(InvoicePrintServiceImpl.class).print(invoicingProject.getInvoice(), null));
+          Beans.get(InvoicePrintServiceImpl.class).print(invoicingProject.getInvoice(), null,null));
       fileList.add(reportSettings.generate().getFile());
 
       MetaFile metaFile = metaFiles.upload(PdfTool.mergePdf(fileList));